[geometry] Move MeshField and SurfaceMesh to proximity. (#11536)
* [geometry] Move MeshField and SurfaceMesh to proximity.
