+++ geometry/proximity/BUILD.bazel
@@ -17,6 +17,7 @@ drake_cc_package_library(
         ":distance_to_point",
         ":distance_to_point_with_gradient",
         ":distance_to_shape",
+        ":mesh_field",
         ":proximity_utilities",
         ":volume_mesh",
     ],
@@ -71,6 +72,23 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "mesh_field",
+    srcs = [
+        "mesh_field.cc",
+        "mesh_field_linear.cc",
+        "surface_mesh.cc",
+    ],
+    hdrs = [
+        "mesh_field.h",
+        "mesh_field_linear.h",
+        "surface_mesh.h",
+    ],
+    deps = [
+        "//common",
+    ],
+)
+
+drake_cc_library(
     name = "proximity_utilities",
     srcs = ["proximity_utilities.cc"],
     hdrs = ["proximity_utilities.h"],
@@ -95,7 +113,7 @@ drake_cc_library(
     deps = [
         "//common",
         "//geometry:geometry_ids",
-        "//geometry/query_results:mesh_field",
+        "//geometry/proximity:mesh_field",
     ],
 )
 
@@ -145,14 +163,14 @@ drake_cc_googletest(
 drake_cc_googletest(
     name = "mesh_field_test",
     deps = [
-        "//geometry/query_results:mesh_field",
+        "//geometry/proximity:mesh_field",
     ],
 )
 
 drake_cc_googletest(
     name = "mesh_field_linear_test",
     deps = [
-        "//geometry/query_results:mesh_field",
+        "//geometry/proximity:mesh_field",
     ],
 )
