+++ geometry/proximity/mesh_field_linear.cc
@@ -1,4 +1,4 @@
-#include "drake/geometry/query_results/mesh_field_linear.h"
+#include "drake/geometry/proximity/mesh_field_linear.h"
 
 namespace drake {
 namespace geometry {
