+++ geometry/proximity/test/mesh_field_test.cc
@@ -1,10 +1,10 @@
-#include "drake/geometry/query_results/mesh_field.h"
+#include "drake/geometry/proximity/mesh_field.h"
 
 #include <memory>
 
 #include <gtest/gtest.h>
 
-#include "drake/geometry/query_results/surface_mesh.h"
+#include "drake/geometry/proximity/surface_mesh.h"
 
 namespace drake {
 namespace geometry {
