+++ geometry/query_results/contact_surface.h
@@ -8,8 +8,8 @@
 #include "drake/common/drake_copyable.h"
 #include "drake/common/eigen_types.h"
 #include "drake/geometry/geometry_ids.h"
-#include "drake/geometry/query_results/mesh_field_linear.h"
-#include "drake/geometry/query_results/surface_mesh.h"
+#include "drake/geometry/proximity/mesh_field_linear.h"
+#include "drake/geometry/proximity/surface_mesh.h"
 
 namespace drake {
 namespace geometry {
